fix: added provider validation for deleting provider

+++ src/admin/admin.service.js
@@ -150,6 +150,8 @@ export const adminUpdateProviderService = async (req, res, next) => {
 
 export const adminDeleteProviderService = async (req, res, next) => {
   try {
+    const existingProvider = await findProviderById(req.body.provider);
+    if (!existingProvider) throw new BaseHttpError(617);
     const soldCoupons = await getProviderSoldCoupons(req.body.provider);
     if (soldCoupons.length !== 0) throw new BaseHttpError(628);
     await deleteProviderCouponsRepository(req.body.provider);
